leetcode/random/longestSubArray1sAfterDeleting/longestSubArray_test.go
package longestsubarray1safterdeleting

import "testing"

func Test_longestSubarray(t *testing.T) {
	type args struct {
		nums []int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			name: "Case 1",
			args: args{
				nums: []int{1, 1, 0, 1},
			},
			want: 3,
		},
		{
			name: "Case 2",
			args: args{
				nums: []int{0, 1, 1, 1, 0, 1, 1, 0, 1},
			},
			want: 5,
		},
		{
			name: "Case 3",
			args: args{
				nums: []int{1, 1, 1},
			},
			want: 2,
		},
		{
			name: "Case 4",
			args: args{
				nums: []int{1, 1, 0, 0, 1, 1, 1, 0, 1},
			},
			want: 4,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := longestSubarray(tt.args.nums); got != tt.want {
				t.Errorf("longestSubarray() = %v, want %v", got, tt.want)
			}
		})
	}
}
